pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jack-o-linter"
version = "0.1.0"
description = "All-in-one static and dynamic analysis tool for Python source code."
readme = "README.md"
authors = [
  {name = "Jack Swiney", email = "jack.swiney11@gmail.com"}
]
maintainers = [
  {name = "Jack Swiney", email = "jack.swiney11@gmail.com"}
]
classifiers = [

]
license = {text = "Apache Software License 2.0"}
dependencies = [
    "flake8",
    "mypy",
    "pydoclint",
    "pygount",
    "pylint",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "pylint",
    "pytest"
]

[project.urls]

bugs = "https://github.com/jack-swiney/jack-o-linter/issues"
changelog = "https://github.com/jack-swiney/jack-o-linter/blob/master/changelog.md"
homepage = "https://github.com/jack-swiney/jack-o-linter"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.coverage.run]
command_line = "-m pytest"
source = [
  "src/jack_o_linter",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__:",
]